check if messageQueue[symbol] is an array
Got some errors when fetching the market depth for all symbols. Turns out that the messageQueue is not initiated with an empty array for each symbol. 

It makes sense to first check it the messageQueue for the symbol is set. If not, then it falls back to the depthhandler function.

I tested it locally, error is gone, all bids and asks are getting calculated correctly. Hope this helps to make this wonderful package more stable guys. Cheers.