Fix endless script
`setInterval` prevents any script from finishing because it keeps pushing
work to the event loop. As soon as you require `node-binance-api`, the
`setInterval` is started regardless you have or not open websockets.

Start/Set the interval when the first socket is open.
Stop/Clear the interval when all sockets are closed.