Update node-binance-api.js
Hi,

Fixed these errors when checking the marketDepth stream:

'CombinedStream: Parse error: Cannot read property \'push\' of undefined' 

I added a check to see if `context.messageQueue` is there. If not, it falls back to the depthHandler function. 

Only tested this locally, but errors went away and worked like a charm again!