Relax depthCache out of sync check
It is possible that the exchange will send a range of depth updates, a subset, at the start, of which we have already cached. This should be allowed by the out of sync check.

The out of sync check will now only treat a gap between the cached data, and newly received data as truly out of sync.

Thanks @learnathoner